+++ crates/squawk_parser/test_data/ok/create_view_extra_parens.sql
@@ -1,0 +1,3 @@
+create view foo as 
+  select b.y from (( select y from bar )) as b
+  order by y desc;